+++ src/Pim/Component/Connector/Reader/DummyItemReader.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Pim\Bundle\BaseConnectorBundle\Reader;
+namespace Pim\Component\Connector\Reader;
 
 use Akeneo\Bundle\BatchBundle\Item\AbstractConfigurableStepElement;
 use Akeneo\Bundle\BatchBundle\Item\ItemReaderInterface;
@@ -9,12 +9,10 @@ use Akeneo\Bundle\BatchBundle\Item\ItemReaderInterface;
  * Dummy step, can't be used unless you have a concrete implementation
  *
  * @author    Nicolas Dupont <nicolas@akeneo.com>
- * @copyright 2013 Akeneo SAS (http://www.akeneo.com)
+ * @copyright 2015 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
- *
- * @deprecated please use Pim\Component\Connector\Reader\DummyItemReader
  */
-class DummyReader extends AbstractConfigurableStepElement implements ItemReaderInterface
+class DummyItemReader implements ItemReaderInterface
 {
     /**
      * {@inheritdoc}
@@ -22,13 +20,5 @@ class DummyReader extends AbstractConfigurableStepElement implements ItemReaderI
     public function read()
     {
         return null;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getConfigurationFields()
-    {
-        return array();
     }
 }